Using a common card component
