+++ src/Components/Admin/Rooms/rooms.js
@@ -3,8 +3,9 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 import Container from "@material-ui/core/Container";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
-import RoomCard from "./roomcard.js";
 import Grid from "@material-ui/core/Grid";
+
+import ItemCard from "../common/itemCard.js";
 
 const useStyles = makeStyles((theme) => ({
     page: {
@@ -39,11 +40,12 @@ export default function EscapeRooms({
                 >
                     {environments.map(function (room) {
                         return (
-                            <RoomCard
+                            <ItemCard
                                 key={room.id}
                                 data={room}
-                                handleEditRoomClick={handleEditRoomClick}
-                                handleDeleteRoomClick={handleDeleteRoomClick}
+                                cardType="environment"
+                                handleEditClick={handleEditRoomClick}
+                                handleDeleteClick={handleDeleteRoomClick}
                             />
                         );
                     })}
